build: create own bundle for `@angular/compiler`
With this change we create a a package to bundle `@angular/compiler`. This is in preparation for version 13 were the angular packages are only shipped in ESM format however due to limitations in ESLint these cannot be consumed directly. Therefore, we have created a "private" `@angular-eslint/bundled-angular-compiler` package which contains the a CJS bundled version of the `@angular/compiler`.  All references to `@angular/compiler` in the code base have been changed to reference this change.