fix: move typescript-eslint packages to peerDeps, consistently allow v7 and v8 (#1856)
As part of the v18 release, support for typescript-eslint v8 prereleases was added. For authors of custom rules that consume the `@angular-eslint/utils` or new `@angular-eslint/test-utils` packages, the initial way this was set up with dependencies could cause type conflicts between v7 and v8. To resolve this the `@angular-eslint/*` packages now depend on `@typescript-eslint/*` packages via a peerDependency which permits both v7 and v8 prereleases.