feat(eslint-plugin-template): Add `no-nested-tags` rule
denies nesting a and p tags

If give a browser nested a or p tags, it'll de-nest them. It's possible via `appendChild` though. So if you never used Ssr and use angular only browser you may never notices.

But now with the addition of https://angular.dev/guide/incremental-hydration this is a problem. If you have a component that creates:

`<a>  <div>  <a>inner</a>  </div>  </a>`

Then Ssr will send:

`<a>  </a><div><a>  </a><a>inner</a>  </div>`

Whis is a DOM mismatch and incremental-hydration will throw an error and not successfully bootstrap.

This lint rule tries to prevent this. Of course not across components, but at least if it happens inside a single template.

related: https://github.com/angular/angular/issues/51399